Moves online version check to separate function.
By default, the ipwb version check makes a request to pypi.org to determine if the
currently installed version is current. This should not occur -- the command should
be executed locally and the current version reported. However, this functionality is
useful, so I have put it behind a separate flag.

Closes #676