Fix anti-aliasing of logo in README
The logo in the README recently was adapted to be more visible
in dark mode (#689), but this caused the smaller text near the
bottom to be more pixely in light mode. After some evaluation
of the source PSD, I noticed that the bottom text was the only
layer that had strong anti-aliasing while the others were sharp.

The small text being sharp is more noticable in the bottom text
than the larger. This PR adds a separate PSD for editing the
logo with the large stroke so the scaling procedure does not
have to be performed on the high-res basis.