Add dark mode CSS, improve accessibility.
Adds an initial set of style for the replay UI to be adaptive to
dark mode. This also moves to use the version of the logo with
the light border that is only apparent in dark mode. The kerning,
coincidentally, is better in this version than the version that
has been used in the replay UI.

Some other tweaks have also been applied to improve the
accessibility including some markup additions and adjustments.

This PR consists of just some initial style for dark mode, and
could use some further tweaking, but it should meet the requirements
of #695 in making the UI slightly adaptive.

Closes #695