An approach for #760, see below for caveats
This approach allows the replay to run on the specified port but
the Link header for mementos does not use this value. It instead
uses the value in the ipwb config. However, we might not want to
overwrite the port in the config for this one-off specification.
Let's see which way @ibnesayeed prefers in #760 before we move
forward in remedying the issue.